requirements.txt
hmmlearn>=0.3.0
yfinance>=0.2.36
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
ta>=0.11.0
python-telegram-bot>=20.0
python-dotenv>=1.0
requests>=2.31
pytest>=7.4
joblib>=1.3
